Refactored dialogs to their own directory
