+++ Spuds-Ski-Map/src/main/res/layout-land/live_map_options.xml
@@ -15,7 +15,7 @@
         android:paddingHorizontal="@dimen/menu_edge_margin"
         android:layout_marginTop="@dimen/menu_image_top_margin">
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_chairlift"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -27,7 +27,7 @@
             app:enabled_menu_icon="@drawable/ic_chairlift"
             app:enabled_menu_title="@string/show_chairlifts" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_green_runs"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -39,7 +39,7 @@
             app:enabled_menu_icon="@drawable/ic_green"
             app:enabled_menu_title="@string/show_green" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_blue_runs"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -51,7 +51,7 @@
             app:enabled_menu_icon="@drawable/ic_blue"
             app:enabled_menu_title="@string/show_blue" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_black_runs"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -63,7 +63,7 @@
             app:enabled_menu_icon="@drawable/ic_black"
             app:enabled_menu_title="@string/show_black" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_double_black_runs"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -87,7 +87,7 @@
         android:layout_marginTop="@dimen/menu_image_top_margin"
         android:layout_marginBottom="@dimen/menu_image_top_margin">
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_terrain_parks"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -99,7 +99,7 @@
             app:enabled_menu_icon="@drawable/ic_terrain_park"
             app:enabled_menu_title="@string/show_terrain" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_night_runs"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -111,7 +111,7 @@
             app:disabled_menu_icon="@drawable/ic_sun"
             app:disabled_menu_title="@string/all_runs" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/toggle_location_tracking"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
