+++ Spuds-Ski-Map/src/main/res/layout/info_map_options.xml
@@ -15,7 +15,7 @@
         android:paddingHorizontal="@dimen/menu_edge_margin"
         android:layout_marginTop="@dimen/menu_image_top_margin">
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_chairlift"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -27,7 +27,7 @@
             app:disabled_menu_icon="@drawable/ic_chairlift_disabled"
             app:disabled_menu_title="@string/hide_chairlifts" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_green_runs"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -39,7 +39,7 @@
             app:disabled_menu_icon="@drawable/ic_green_disabled"
             app:disabled_menu_title="@string/hide_green" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_blue_runs"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -61,7 +61,7 @@
         android:paddingHorizontal="@dimen/menu_edge_margin"
         android:layout_marginTop="@dimen/menu_image_top_margin">
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_black_runs"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -73,7 +73,7 @@
             app:disabled_menu_icon="@drawable/ic_black_disabled"
             app:disabled_menu_title="@string/hide_black" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_double_black_runs"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -85,7 +85,7 @@
             app:enabled_menu_icon="@drawable/ic_missing"
             app:enabled_menu_title="@string/show_double_black" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_terrain_parks"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -109,7 +109,7 @@
         android:layout_marginTop="@dimen/menu_image_top_margin"
         android:layout_marginBottom="@dimen/menu_image_top_margin">
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_night_runs"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
@@ -121,7 +121,7 @@
             app:disabled_menu_icon="@drawable/ic_sun"
             app:disabled_menu_title="@string/all_runs" />
 
-        <xyz.thespud.skimap.activities.MapOptionItem
+        <xyz.thespud.skimap.dialogs.MapOptionItem
             android:id="@+id/show_circles"
             android:layout_width="@dimen/menu_image_size"
             android:layout_height="wrap_content"
