+++ Spuds-Ski-Map/src/main/res/layout/info_map_options_v2.xml
@@ -8,7 +8,7 @@
     android:useDefaultMargins="true"
     android:padding="@dimen/menu_edge_margin">
 
-    <xyz.thespud.skimap.activities.MapOptionItem
+    <xyz.thespud.skimap.dialogs.MapOptionItem
         android:id="@+id/show_chairlift"
         android:layout_width="@dimen/menu_image_size"
         android:layout_height="wrap_content"
@@ -22,7 +22,7 @@
         app:disabled_menu_title="@string/hide_chairlifts" />
 
 
-    <xyz.thespud.skimap.activities.MapOptionItem
+    <xyz.thespud.skimap.dialogs.MapOptionItem
         android:id="@+id/show_green_runs"
         android:layout_width="@dimen/menu_image_size"
         android:layout_height="wrap_content"
@@ -35,7 +35,7 @@
         app:disabled_menu_icon="@drawable/ic_green_disabled"
         app:disabled_menu_title="@string/hide_green" />
 
-    <xyz.thespud.skimap.activities.MapOptionItem
+    <xyz.thespud.skimap.dialogs.MapOptionItem
         android:id="@+id/show_blue_runs"
         android:layout_width="@dimen/menu_image_size"
         android:layout_height="wrap_content"
@@ -48,7 +48,7 @@
         app:disabled_menu_icon="@drawable/ic_blue_disabled"
         app:disabled_menu_title="@string/hide_blue" />
 
-    <xyz.thespud.skimap.activities.MapOptionItem
+    <xyz.thespud.skimap.dialogs.MapOptionItem
         android:id="@+id/show_black_runs"
         android:layout_width="@dimen/menu_image_size"
         android:layout_height="wrap_content"
@@ -61,7 +61,7 @@
         app:disabled_menu_icon="@drawable/ic_black_disabled"
         app:disabled_menu_title="@string/hide_black" />
 
-    <xyz.thespud.skimap.activities.MapOptionItem
+    <xyz.thespud.skimap.dialogs.MapOptionItem
         android:id="@+id/show_double_black_runs"
         android:layout_width="@dimen/menu_image_size"
         android:layout_height="wrap_content"
@@ -74,7 +74,7 @@
         app:enabled_menu_icon="@drawable/ic_missing"
         app:enabled_menu_title="@string/show_double_black" />
 
-    <xyz.thespud.skimap.activities.MapOptionItem
+    <xyz.thespud.skimap.dialogs.MapOptionItem
         android:id="@+id/show_terrain_parks"
         android:layout_width="@dimen/menu_image_size"
         android:layout_height="wrap_content"
@@ -87,7 +87,7 @@
         app:enabled_menu_icon="@drawable/ic_terrain_park"
         app:enabled_menu_title="@string/show_terrain" />
 
-    <xyz.thespud.skimap.activities.MapOptionItem
+    <xyz.thespud.skimap.dialogs.MapOptionItem
         android:id="@+id/show_night_runs"
         android:layout_width="@dimen/menu_image_size"
         android:layout_height="wrap_content"
@@ -100,7 +100,7 @@
         app:disabled_menu_icon="@drawable/ic_sun"
         app:disabled_menu_title="@string/all_runs" />
 
-    <xyz.thespud.skimap.activities.MapOptionItem
+    <xyz.thespud.skimap.dialogs.MapOptionItem
         android:id="@+id/show_circles"
         android:layout_width="@dimen/menu_image_size"
         android:layout_height="wrap_content"
